ai-engine/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
attrs==18.2.0
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.2.22
langchain-openai==1.1.12
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
lxml==6.0.2
openai==2.29.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pdflatex==0.1.3
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.2.2
python-docx==1.2.0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.1.0